+++ R/create.R
@@ -106,13 +106,12 @@ create_project <- function(path,
 
 #' Create a project from a GitHub repo
 #'
-#' @description
-#' Creates a new local project and Git repository from a repo on GitHub, by
-#' either cloning or
-#' [fork-and-cloning](https://docs.github.com/articles/fork-a-repo). In the
-#' fork-and-clone case, `create_from_github()` also does additional remote and
-#' branch setup, leaving you in the perfect position to make a pull request with
-#' [pr_init()], one of several [functions that work pull
+#' @description Creates a new local project and Git repository from a repo on
+#' GitHub, by either cloning or
+#' [fork-and-cloning](https://docs.github.com/en/get-started/quickstart/fork-a-repo).
+#' In the fork-and-clone case, `create_from_github()` also does additional
+#' remote and branch setup, leaving you in the perfect position to make a pull
+#' request with [pr_init()], one of several [functions that work pull
 #' requests][pull-requests].
 #'
 #' `create_from_github()` works best when your GitHub credentials are
@@ -224,7 +223,7 @@ create_from_github <- function(repo_spec,
 
   source_owner <- spec_owner(repo_spec)
   repo_name <- spec_repo(repo_spec)
-  gh <- gh_tr(list(repo_owner = source_owner, repo_name = repo_name, .api_url = host))
+  gh <- gh_tr(list(repo_owner = source_owner, repo_name = repo_name, api_url = host))
 
   repo_info <- gh("GET /repos/{owner}/{repo}")
   # 2020-10-14 GitHub has had some bugs lately around default branch
@@ -329,7 +328,7 @@ challenge_nested_project <- function(path, name) {
     is regarded as a project."
   )
   if (ui_nope("Do you want to create anyway?")) {
-    ui_stop("Aborting project creation.")
+    ui_stop("Cancelling project creation.")
   }
   invisible()
 }