+++ R/logo.R
@@ -28,7 +28,7 @@ use_logo <- function(img, geometry = "240x278", retina = TRUE) {
 
   if (path_ext(img) == "svg") {
     logo_path <- path("man", "figures", "logo.svg")
-    file_copy(img, proj_path(logo_path))
+    file_copy(img, proj_path(logo_path), overwrite = TRUE)
     ui_done("Copied {ui_path(img)} to {ui_path(logo_path)}")
 
     height <- as.integer(sub(".*x", "", geometry))