+++ R/pkgdown.R
@@ -26,8 +26,9 @@
 #' Travis-CI.
 #'
 #' @seealso <https://pkgdown.r-lib.org/articles/pkgdown.html#configuration>
-#' @param config_file Path to the pkgdown yaml config file
-#' @param destdir Target directory for pkgdown docs
+#' @param config_file Path to the pkgdown yaml config file, relative to the
+#'  project.
+#' @param destdir Target directory for pkgdown docs.
 #' @export
 use_pkgdown <- function(config_file = "_pkgdown.yml", destdir = "docs") {
   check_is_package("use_pkgdown()")
@@ -68,7 +69,8 @@ pkgdown_version <- function() {
 #' @rdname use_pkgdown
 #' @export
 use_pkgdown_github_pages <- function() {
-  tr <- target_repo(github_get = TRUE)
+  tr <- target_repo(github_get = TRUE, ok_configs = c("ours", "fork"))
+  check_can_push(tr = tr, "to turn on GitHub Pages")
 
   use_pkgdown()
   site <- use_github_pages()
@@ -77,7 +79,7 @@ use_pkgdown_github_pages <- function() {
   site_url <- tidyverse_url(url = site$html_url, tr = tr)
   use_pkgdown_url(url = site_url, tr = tr)
 
-  if (tr$repo_owner %in% c("tidyverse", "tidymodels", "r-lib")) {
+  if (is_rstudio_pkg()) {
     ui_done("
       Adding {ui_value('tidyverse/tidytemplate')} to \\
       {ui_field('Config/Needs/website')}")
@@ -105,6 +107,10 @@ use_pkgdown_url <- function(url, tr = NULL) {
   desc <- desc::desc(file = proj_get())
   desc$add_urls(url)
   desc$write()
+  if (has_package_doc()) {
+    ui_todo("
+      Run {ui_code('devtools::document()')} to update package-level documentation.")
+  }
 
   gh <- gh_tr(tr)
   homepage <- gh("GET /repos/{owner}/{repo}")[["homepage"]]